Tweak About view, make links actually clickable
(cherry picked from commit 68286b2accbdfe6fb111f17f38311e41bd4dfdb7)
