+++ app/src/main/java/eu/kanade/tachiyomi/ui/more/MoreHeaderPreference.kt
@@ -1,0 +1,15 @@
+package eu.kanade.tachiyomi.ui.more
+
+import android.content.Context
+import android.util.AttributeSet
+import androidx.preference.Preference
+import eu.kanade.tachiyomi.R
+
+class MoreHeaderPreference @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null) :
+    Preference(context, attrs) {
+
+    init {
+        layoutResource = R.layout.pref_more_header
+        isSelectable = false
+    }
+}
